chore: integrate native access with current main
Preserve the 3.8 platform-auth and model-routing additions while rebasing the native-access candidate, add the required patch changeset, and split merged setup/model owners to keep the readability ratchet moving down.
