+++ packages/cli/src/control-services.ts
@@ -46,11 +46,8 @@ import {
 } from "./credential-status-invalidation.js";
 import { createSetupJobManager } from "./setup-jobs.js";
 import { SetupJobStore } from "./setup-job-store.js";
-import { activeProfileLoginJob, preflightSetupJobCreateRequest } from "./setup-job-support.js";
-import {
-  assertSetupLoginAdmission,
-  projectSetupLoginCapability,
-} from "./setup-login-capability.js";
+import { activeProfileLoginJob } from "./setup-job-support.js";
+import { setupJobControlServices } from "./setup-job-control-services.js";
 import { SetupLifecycleBinding } from "./setup-lifecycle-binding.js";
 import { createRunRequirementsPreflight } from "./request-preflight.js";
 import { threadRunStartRequiresGit } from "./thread-execution-workspace.js";
@@ -328,35 +325,7 @@ export function controlServices(
     agentCapabilities: async () => buildAgentCapabilityCatalog(),
     runApplicability: async (input: { repoRoot: string }) =>
       projectRunApplicability(input.repoRoot),
-    createSetupJob: async (input: {
-      request: unknown;
-      idempotencyKey: string;
-      clientId: string;
-    }) => {
-      // No helper/vendor probe and no bootstrap/durable mutation can precede
-      // request, named-profile, required-profile, and D13 validation.
-      const request = preflightSetupJobCreateRequest(input.request);
-      const capability = await projectSetupLoginCapability(request.harness, {
-        transport: request.transport,
-        loginFlow: request.loginFlow,
-      });
-      assertSetupLoginAdmission(capability);
-      return setupJobs().create(request, {
-        key: input.idempotencyKey,
-        client: input.clientId,
-      });
-    },
-    listSetupJobs: async (input?: unknown) => {
-      const jobs = setupJobs();
-      return { jobs: jobs.list(input as Parameters<typeof jobs.list>[0]) };
-    },
-    setupJobStatus: async (input: unknown) => setupJobs().status(input),
-    setupJobSnapshot: async (input: unknown) => setupJobs().snapshot(input),
-    setupJobEvents: async (input: unknown) => setupJobs().events(input),
-    cancelSetupJob: async (input: unknown) => setupJobs().cancel(input),
-    setupJobInput: async (input: unknown) => setupJobs().input(input),
-    reconcileSetupJob: async (input: unknown) => setupJobs().reconcile(input),
-    extendSetupJob: async (input: unknown) => setupJobs().extend(input),
+    ...setupJobControlServices(setupJobs),
     journalEvents: async (partition: string, afterCursor?: string) =>
       journalPartition(partition).events(afterCursor),
     recoveryInspectPartition: async (partition: string) => journalPartition(partition).inspect(),
